I make some code refactoring when deleting records.
The def destroy in the controller now can send if the record wose deleted and if not you can output a message.

Example:

<pre><code class="ruby">
  # products_controller.rb
  def destroy
    product = Product.find(params[:id])
    if product.destroy
      render :json => { :success => true } 
    else
      render :json => { :success => false, :msg => "#{product.name} have some contacts associated" }
    end
  end

  #In product.rb
  def before_destroy
      return contacts.size == 0
  end
</code></pre>

list.html.erb now is a little cleaner,

I removed mask() and unmask() function because are dry, and now we don't need a listener on ds remove. 

The function remove() now can detect if an operation have "succes" or  not an can output the msg that you can set in the controller!



git-svn-id: http://svn.lipsiasoft.com/lipsiadmin@27 9c8155e6-a184-41b9-ba3d-52e82f7d7c69